unify upgrade tests
Unify upgrade images into a single test, extract image name and image
version into vars to be reused.

This will also help with CI as we cna only run 5 macos instances in
parallel so reducing the number of workers to run the tests will speed
up the full workflow
