+++ .github/build.yaml.gomplate
@@ -282,7 +282,7 @@
     needs: vbox-{{{$subset}}}-{{{ $flavor }}}
     strategy:
       matrix:
-        test: ["test-features", "test-smoke", "test-fallback", "test-recovery", "test-upgrades-images-signed", "test-upgrades-images-unsigned", "test-upgrades-local", "test-deploys-images"]
+        test: ["test-features", "test-smoke", "test-fallback", "test-recovery", "test-upgrades-images", "test-upgrades-local", "test-deploys-images"]
     steps:
       - name: Install Go
         uses: actions/setup-go@v2
