--- tests/upgrades-images-unsigned/tests_suite_test.go
@@ -1,13 +1,0 @@
-package cos_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestTests(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "cOS Upgrate test Suite - Images unsigned")
-}
